Start implemting elevator network
